Create Backup2.SLOTS
In order for others to use this, you need to PIP install colorama, and ensure your system has json working. Your sounds wont work correctly because they're not included with the file, as well as you need to create a savedata.txt in the path with 14 variables in a list. Here's one you can copy. [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]